refactor job manager
